requirements.txt
tkcalendar
IMDbPY
XlsxWriter
mysql-connector-python
requests
xlrd
datetime
